Use DataType enum
